Unify Subscription implementations in a single file
